+++ front/src/Search-page/main2.tsx
