+++ front/src/main-page/index.css
