+++ front/src/main-page/local_data.ts
