+++ front/src/main-page/msg.css
