+++ front/src/main-page/msg.tsx
