--- front/test1/src/Search-page/addData.ts
@@ -1,7 +1,0 @@
-import { DBdata } from "../../../../public/type"
-
-export let schoolData: DBdata = {
-    schoolname: "",
-    year: 0,
-    class: 0
-}
