--- front/test1/src/Search-page/local_data.ts
@@ -1,8 +1,0 @@
-import { DBdata } from "../../../../public/type"
-
-export const addLocalData = (data:DBdata) => {
-    const d = new Date();
-    d.setMonth(d.getMonth() + 1)
-    localStorage.setItem('obj', JSON.stringify(data))
-    document.cookie = `id=${data.schoolname}; expires=${d.toUTCString()};`;
-}
