+++ front/vite.config.ts
